services/ml-scoring/requirements.txt
fastapi
uvicorn
torch
transformers
Pillow
python-multipart
sentencepiece
protobuf
